feat: Add "from" field to email options
The code changes add a new "from" field to the email options in the `/send-email` route. This allows specifying the sender name in addition to the email address. The value of the "from" field is set to either the decoded "from" value from the token payload or the default value "Mailer API". If the "to" field is missing in the token payload, a 400 response with the message "Invalid Token. Email not found." is returned.